move to pkg/trace
