+++ controller/cmd/identity/main.go
@@ -19,7 +19,7 @@ import (
 	consts "github.com/linkerd/linkerd2/pkg/k8s"
 	"github.com/linkerd/linkerd2/pkg/prometheus"
 	"github.com/linkerd/linkerd2/pkg/tls"
-	"github.com/linkerd/linkerd2/pkg/util"
+	"github.com/linkerd/linkerd2/pkg/trace"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -122,7 +122,7 @@ func Main(args []string) {
 	}
 
 	if *traceCollector != "" {
-		if err := util.InitializeTracing("linkerd-identity", *traceCollector); err != nil {
+		if err := trace.InitializeTracing("linkerd-identity", *traceCollector); err != nil {
 			log.Warnf("failed to initialize tracing: %s", err)
 		}
 	}
