+++ controller/cmd/tap/main.go
@@ -14,7 +14,7 @@ import (
 	"github.com/linkerd/linkerd2/pkg/config"
 	"github.com/linkerd/linkerd2/pkg/flags"
 	pkgK8s "github.com/linkerd/linkerd2/pkg/k8s"
-	"github.com/linkerd/linkerd2/pkg/util"
+	"github.com/linkerd/linkerd2/pkg/trace"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -65,7 +65,7 @@ func Main(args []string) {
 	log.Info("Using cluster domain: ", clusterDomain)
 
 	if *traceCollector != "" {
-		if err := util.InitializeTracing("linkerd-tap", *traceCollector); err != nil {
+		if err := trace.InitializeTracing("linkerd-tap", *traceCollector); err != nil {
 			log.Warnf("failed to initialize tracing: %s", err)
 		}
 	}
